requirements.txt
# Install the SDK and dependencies
PowerPlatform-Dataverse-Client
azure-identity
langchain>=1.0.0
langgraph>=1.0.0
langchain-openai
langchain-community
openai
python-dotenv
pandas
plotly
